deler opp pakkene i behandling og vedtak
