+++ src/main/kotlin/no/nav/tiltakspenger/datadeling/Application.kt
@@ -22,18 +22,18 @@ import no.nav.tiltakspenger.datadeling.application.db.DataSourceSetup
 import no.nav.tiltakspenger.datadeling.application.exception.ExceptionHandler
 import no.nav.tiltakspenger.datadeling.application.routes.healthRoutes
 import no.nav.tiltakspenger.datadeling.application.routes.swaggerRoute
+import no.nav.tiltakspenger.datadeling.behandling.datadeling.BehandlingService
+import no.nav.tiltakspenger.datadeling.behandling.datadeling.routes.behandlingRoutes
+import no.nav.tiltakspenger.datadeling.behandling.db.BehandlingRepo
+import no.nav.tiltakspenger.datadeling.behandling.motta.MottaNyBehandlingService
 import no.nav.tiltakspenger.datadeling.client.arena.ArenaClient
 import no.nav.tiltakspenger.datadeling.identhendelse.IdenthendelseConsumer
 import no.nav.tiltakspenger.datadeling.identhendelse.IdenthendelseService
-import no.nav.tiltakspenger.datadeling.motta.behandling.MottaNyBehandlingService
-import no.nav.tiltakspenger.datadeling.motta.behandling.db.BehandlingRepo
 import no.nav.tiltakspenger.datadeling.motta.mottaRoutes
-import no.nav.tiltakspenger.datadeling.motta.vedtak.MottaNyttVedtakService
-import no.nav.tiltakspenger.datadeling.motta.vedtak.db.VedtakRepo
-import no.nav.tiltakspenger.datadeling.routes.behandling.behandlingRoutes
-import no.nav.tiltakspenger.datadeling.routes.vedtak.vedtakRoutes
-import no.nav.tiltakspenger.datadeling.service.BehandlingService
-import no.nav.tiltakspenger.datadeling.service.VedtakService
+import no.nav.tiltakspenger.datadeling.vedtak.datadeling.VedtakService
+import no.nav.tiltakspenger.datadeling.vedtak.datadeling.routes.vedtakRoutes
+import no.nav.tiltakspenger.datadeling.vedtak.db.VedtakRepo
+import no.nav.tiltakspenger.datadeling.vedtak.motta.MottaNyttVedtakService
 import no.nav.tiltakspenger.libs.common.GenerellSystembruker
 import no.nav.tiltakspenger.libs.common.GenerellSystembrukerrolle
 import no.nav.tiltakspenger.libs.common.GenerellSystembrukerroller
