+++ src/main/kotlin/no/nav/tiltakspenger/datadeling/motta/MottaRoutes.kt
@@ -1,10 +1,10 @@
 package no.nav.tiltakspenger.datadeling.motta
 
 import io.ktor.server.routing.Route
-import no.nav.tiltakspenger.datadeling.motta.behandling.MottaNyBehandlingService
-import no.nav.tiltakspenger.datadeling.motta.behandling.route.mottaNyBehandlingRoute
-import no.nav.tiltakspenger.datadeling.motta.vedtak.MottaNyttVedtakService
-import no.nav.tiltakspenger.datadeling.motta.vedtak.route.mottaNyttVedtakRoute
+import no.nav.tiltakspenger.datadeling.behandling.motta.MottaNyBehandlingService
+import no.nav.tiltakspenger.datadeling.behandling.motta.routes.mottaNyBehandlingRoute
+import no.nav.tiltakspenger.datadeling.vedtak.motta.MottaNyttVedtakService
+import no.nav.tiltakspenger.datadeling.vedtak.motta.routes.mottaNyttVedtakRoute
 import java.time.Clock
 
 fun Route.mottaRoutes(
