--- src/main/kotlin/no/nav/tiltakspenger/datadeling/routes/behandling/BehandlingResponse.kt
@@ -1,24 +1,0 @@
-package no.nav.tiltakspenger.datadeling.routes.behandling
-
-import no.nav.tiltakspenger.datadeling.domene.Behandling
-import java.time.LocalDate
-
-/**
- * Kontrakt for åpne behandlinger (filtrerer bort avsluttede behandlinger: iverksatt+avbrutt).
- */
-data class BehandlingResponse(
-    val behandlingId: String,
-    val fom: LocalDate,
-    val tom: LocalDate,
-)
-
-internal fun List<Behandling>.toResponse(): List<BehandlingResponse> {
-    return this.map { it.toBehandlingResponse() }
-}
-
-internal fun Behandling.toBehandlingResponse() =
-    BehandlingResponse(
-        behandlingId = this.behandlingId,
-        fom = this.periode.fraOgMed,
-        tom = this.periode.tilOgMed,
-    )
