+++ src/main/kotlin/no/nav/tiltakspenger/datadeling/vedtak/motta/MottaNyttVedtakService.kt
@@ -1,0 +1,21 @@
+package no.nav.tiltakspenger.datadeling.vedtak.motta
+
+import arrow.core.Either
+import no.nav.tiltakspenger.datadeling.domene.TiltakspengerVedtak
+import no.nav.tiltakspenger.datadeling.vedtak.db.VedtakRepo
+
+class MottaNyttVedtakService(
+    private val vedtakRepo: VedtakRepo,
+) {
+    fun motta(
+        vedtak: TiltakspengerVedtak,
+    ): Either<KanIkkeMottaVedtak, Unit> {
+        return Either.catch { vedtakRepo.lagre(vedtak) }.mapLeft {
+            KanIkkeMottaVedtak.Persisteringsfeil
+        }
+    }
+}
+
+sealed interface KanIkkeMottaVedtak {
+    data object Persisteringsfeil : KanIkkeMottaVedtak
+}
